requirements.txt
gradio
openai
python-dotenv
pydub
soundfile
